feat: boiler plate macros
