+++ mysk-lib-macros/src/fetch_level_variant/mod.rs
@@ -1,1 +1,1 @@
-pub mod non_db_request_variant;
+pub mod impl_fetch_level_variant_from;
