+++ mysk-lib/src/models/student/fetch_levels/id_only.rs
@@ -1,4 +1,4 @@
-use mysk_lib_macros::fetch_level_variant::non_db_request_variant;
+use mysk_lib_macros::{id_only_variant_boiler_plate, impl_fetch_level_variant_from};
 use serde::{Deserialize, Serialize};
 use sqlx::PgPool;
 use uuid::Uuid;
@@ -8,11 +8,9 @@ use crate::models::{
     student::db::DbStudent,
 };
 
-use mysk_lib_macros::non_db_request_variant;
-
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct IdOnlyStudent {
     pub id: Uuid,
 }
 
-non_db_request_variant!(IdOnlyStudent, DbStudent);
+id_only_variant_boiler_plate!(IdOnlyStudent, DbStudent);
